Add files via upload
Task 1: Mobile App Signup Flow 
A mobile app signup flow is the process of users creating an account and signing up for an
app. The goal of a good signup flow is to make it easy and quick for users to sign up, while
also collecting the necessary information from them.