Add files via upload
Restaurant menu
A restaurant menu UI/UX project is the process of designing a menu that is
both visually appealing and easy to use. The goal of a good restaurant menu
is to make it easy for customers to find the dishes they want, and to
encourage them to order more.