Add files via upload
TASK 1 Mobile App Signup Flow A mobile app signup flow is the process of users creating an account and signing up for an app. Goal: The goal of a good signup flow is to make it easy and quick for users to sign up. Personalized: The more personalized the signup flow is, the more likely users are to complete it. I personalize the experience by using the user's name, location, or interests. Useful: The signup flow  collect the necessary information from usersError-free: The signup flow is free of errors, so that users do not have to repeat the process. Secure: The signup flow is secure, so that users' personal information is protected.

Figma Link: https://www.figma.com/file/MZgqWhvfRjnsCw5SAlEdV2/Codsoft-task-1-Mobile-app-sign-up?type=design&node-id=0%3A1&mode=design&t=0Qm7BbYMplzZevog-1